+++ test/unit/transformations/spark_transformations/test_nan.py
@@ -19,6 +19,7 @@ from tmlt.core.metrics import (
     AbsoluteDifference,
     HammingDistance,
     IfGroupedBy,
+    RootSumOfSquared,
     SumOf,
     SymmetricDifference,
 )
@@ -128,8 +129,8 @@ class TestDropInfs(PySparkTest):
             ("At least one column must be specified", []),
             (re.escape("`columns` must not contain duplicate names"), ["B", "B"]),
             (
-                "Inner metric for IfGroupedBy metric must be L1 or L2 over"
-                " SymmetricDifference",
+                "Inner metric for IfGroupedBy metric must be SymmetricDifference, "
+                "or L1 or L2 over SymmetricDifference.",
                 ["B"],
                 IfGroupedBy("A", SumOf(AbsoluteDifference())),
             ),
@@ -148,7 +149,11 @@ class TestDropInfs(PySparkTest):
             )
 
     @parameterized.expand(
-        [(SymmetricDifference(),), (IfGroupedBy("A", SumOf(SymmetricDifference())),)]
+        [
+            (SymmetricDifference(),),
+            (IfGroupedBy("A", SumOf(SymmetricDifference())),),
+            (IfGroupedBy("A", SymmetricDifference()),),
+        ]
     )
     def test_stability_function(
         self, input_metric: Union[SymmetricDifference, IfGroupedBy]
@@ -225,8 +230,8 @@ class TestDropNaNs(PySparkTest):
             ("At least one column must be specified", []),
             ("`columns` must not contain duplicate names", ["B", "B"]),
             (
-                "Inner metric for IfGroupedBy metric must be L1 or L2 over"
-                " SymmetricDifference",
+                "Inner metric for IfGroupedBy metric must be SymmetricDifference, "
+                "or L1 or L2 over SymmetricDifference.",
                 ["B"],
                 IfGroupedBy("A", SumOf(AbsoluteDifference())),
             ),
@@ -245,7 +250,11 @@ class TestDropNaNs(PySparkTest):
             )
 
     @parameterized.expand(
-        [(SymmetricDifference(),), (IfGroupedBy("A", SumOf(SymmetricDifference())),)]
+        [
+            (SymmetricDifference(),),
+            (IfGroupedBy("A", SumOf(SymmetricDifference())),),
+            (IfGroupedBy("A", SymmetricDifference()),),
+        ]
     )
     def test_stability_function(
         self, input_metric: Union[SymmetricDifference, IfGroupedBy]
@@ -330,14 +339,27 @@ class TestDropNulls(PySparkTest):
         self.assertEqual(len(actual_rows), len(expected_rows))
         self.assertEqual(set(actual_rows), set(expected_rows))
 
+    def test_can_drop_grouping_column(self) -> None:
+        """Unlike ReplaceNulls, DropNulls can drop on grouping column."""
+        DropNulls(
+            input_domain=self.input_domain,
+            metric=IfGroupedBy("A", SymmetricDifference()),
+            columns=["A"],
+        )
+        DropNulls(
+            input_domain=self.input_domain,
+            metric=IfGroupedBy("A", RootSumOfSquared(SymmetricDifference())),
+            columns=["A"],
+        )
+
     @parameterized.expand(
         [
             ("One or more columns do not exist in the input domain", ["C"]),
             ("At least one column must be specified", []),
             ("`columns` must not contain duplicate names", ["B", "B"]),
             (
-                "Inner metric for IfGroupedBy metric must be L1 or L2 over"
-                " SymmetricDifference",
+                "Inner metric for IfGroupedBy metric must be SymmetricDifference, "
+                "or L1 or L2 over SymmetricDifference.",
                 ["B"],
                 IfGroupedBy("A", SumOf(AbsoluteDifference())),
             ),
@@ -356,7 +378,11 @@ class TestDropNulls(PySparkTest):
             )
 
     @parameterized.expand(
-        [(SymmetricDifference(),), (IfGroupedBy("A", SumOf(SymmetricDifference())),)]
+        [
+            (SymmetricDifference(),),
+            (IfGroupedBy("A", SumOf(SymmetricDifference())),),
+            (IfGroupedBy("A", SymmetricDifference()),),
+        ]
     )
     def test_stability_function(
         self, input_metric: Union[SymmetricDifference, IfGroupedBy]
@@ -478,6 +504,19 @@ class TestReplaceInfs(PySparkTest):
                 + re.escape("infinity or -infinity"),
                 {"A": (-23, 45)},
             ),
+            (
+                "Inner metric for IfGroupedBy metric must be SymmetricDifference, "
+                "or L1 or L2 over SymmetricDifference.",
+                {"B": (1.0, 23.0)},
+                IfGroupedBy("A", SumOf(AbsoluteDifference())),
+            ),
+            (
+                # if this was supported, it would still not be allowed, because
+                # you can't replace on grouping column. See ReplaceNulls
+                "Can not group by a floating point column: B",
+                {"B": (1.0, 23.0)},
+                IfGroupedBy("B", SumOf(SymmetricDifference())),
+            ),
         ]
     )
     def test_invalid_constructor_args(
@@ -495,7 +534,11 @@ class TestReplaceInfs(PySparkTest):
             )
 
     @parameterized.expand(
-        [(SymmetricDifference(),), (IfGroupedBy("A", SumOf(SymmetricDifference())),)]
+        [
+            (SymmetricDifference(),),
+            (IfGroupedBy("A", SumOf(SymmetricDifference())),),
+            (IfGroupedBy("A", SymmetricDifference()),),
+        ]
     )
     def test_stability_function(
         self, input_metric: Union[SymmetricDifference, IfGroupedBy]
@@ -604,6 +647,19 @@ class TestReplaceNaNs(PySparkTest):
             ("One or more columns do not exist in the input domain", {"C": 0.1}),
             ("At least one column must be specified", {}),
             (r"Replacement value .* is invalid for column \(B\)", {"B": float("nan")}),
+            (
+                "Inner metric for IfGroupedBy metric must be SymmetricDifference, "
+                "or L1 or L2 over SymmetricDifference.",
+                {"B": 1.0},
+                IfGroupedBy("A", SumOf(AbsoluteDifference())),
+            ),
+            (
+                # if this was supported, it would still not be allowed, because
+                # you can't replace on grouping column. See ReplaceNulls
+                "Can not group by a floating point column: B",
+                {"B": 1.0},
+                IfGroupedBy("B", SumOf(SymmetricDifference())),
+            ),
         ]
     )
     def test_invalid_constructor_args(
@@ -612,7 +668,7 @@ class TestReplaceNaNs(PySparkTest):
         replace_map: Dict[str, Any],
         input_metric: Union[SymmetricDifference, IfGroupedBy] = SymmetricDifference(),
     ):
-        """DropNulls raises appropriate errors on invalid constructor arguments."""
+        """ReplaceNaNs raises appropriate errors on invalid constructor arguments."""
         with self.assertRaisesRegex(ValueError, error_msg):
             ReplaceNaNs(
                 input_domain=self.input_domain,
@@ -625,7 +681,7 @@ class TestReplaceNaNs(PySparkTest):
             (SymmetricDifference(),),
             (IfGroupedBy("A", SumOf(SymmetricDifference())),),
             (HammingDistance(),),
-            (IfGroupedBy("A", SumOf(HammingDistance())),),
+            (IfGroupedBy("A", SymmetricDifference()),),
         ]
     )
     def test_stability_function(
@@ -736,6 +792,17 @@ class TestReplaceNulls(PySparkTest):
             ("At least one column must be specified", {}),
             (r"Replacement value .* is invalid for column \(B\)", {"B": None}),
             (r"Replacement value .* is invalid for column \(B\)", {"B": "X"}),
+            (
+                "Inner metric for IfGroupedBy metric must be SymmetricDifference, "
+                "or L1 or L2 over SymmetricDifference.",
+                {"B": 1.0},
+                IfGroupedBy("A", SumOf(AbsoluteDifference())),
+            ),
+            (
+                "Cannot replace values in the grouping column for IfGroupedBy.",
+                {"A": 1},
+                IfGroupedBy("A", SumOf(SymmetricDifference())),
+            ),
         ]
     )
     def test_invalid_constructor_args(
@@ -744,7 +811,7 @@ class TestReplaceNulls(PySparkTest):
         replace_map: Dict[str, Any],
         input_metric: Union[SymmetricDifference, IfGroupedBy] = SymmetricDifference(),
     ):
-        """DropNulls raises appropriate errors on invalid constructor arguments."""
+        """ReplaceNulls raises appropriate errors on invalid constructor arguments."""
         with self.assertRaisesRegex(ValueError, error_msg):
             ReplaceNulls(
                 input_domain=self.input_domain,
@@ -757,7 +824,7 @@ class TestReplaceNulls(PySparkTest):
             (SymmetricDifference(),),
             (IfGroupedBy("A", SumOf(SymmetricDifference())),),
             (HammingDistance(),),
-            (IfGroupedBy("A", SumOf(HammingDistance())),),
+            (IfGroupedBy("A", SymmetricDifference()),),
         ]
     )
     def test_stability_function(